backend/requirements.txt
psycopg2-binary
boto3
requests
pydantic
